requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
tensorboardX>=2.6.0
pyyaml>=5.4.0 